requirements.txt
Flask==2.2.2
google-generativeai==0.3.1
gspread==5.7.2
oauth2client==4.1.3
SpeechRecognition==3.9.0
pydub==0.25.1
gTTS==2.3.0
Werkzeug==2.2.2
Pillow==9.4.0